finish libft

+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: kquispe <kquispe@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/13 12:33:44 by kquispe           #+#    #+#             */
-/*   Updated: 2023/11/01 17:03:12 by kquispe          ###   ########.fr       */
+/*   Updated: 2023/11/17 21:35:29 by kquispe          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
 	unsigned char	*str1;
 	unsigned char	*str2;
-	int				i;
+	size_t			i;
 
 	if (!dest && !src)
 		return (NULL);
