+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: kquispe <kquispe@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/12 12:40:37 by kquispe           #+#    #+#             */
-/*   Updated: 2023/10/17 16:40:04 by kquispe          ###   ########.fr       */
+/*   Updated: 2023/11/17 21:35:38 by kquispe          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 void	*ft_memset(void *ptr, int val, size_t n)
 {
 	unsigned char	*str;
-	int				i;
+	size_t			i;
 
 	i = 0;
 	str = (unsigned char *)ptr;
